Add links to all generated breadcrumbs
This fixes an issue whereby generated breadcrumbs would include SEO
microcode that got picked up as invalid by Google due to invalid IDs
being inferred from them.

For full context, see the issue I posted on Docusaurus:
https://github.com/facebook/docusaurus/issues/7241

A way out of this would be to remove the breadcrumbs, but I opted for
adding a generated index page to each category in the sidebar in order
to give every breadcrumb item a link.